cmake: Add option to build with system jsoncpp.
This patch adds the CMake build option `USE_SYSTEM_JSONCPP`,
which instead of using the internal `jsoncpp.cpp` links against
the system installed jsoncpp library.

This results in a `libopenvr_api.so` that is only 115K instead of 301K
on my system.

Distributions like Debian will prefer this behaviour, since it will give
them the possibility to maintain version and security issues for jsoncpp
in one place.

This behaviour can be enabled by

```
cmake -DUSE_SYSTEM_JSONCPP=True .
```

When the flag is not set, the build will behave like it did before.

I tested this patch using the system wide jsoncpp successfully
with xrdesktop.