[json]: Use default ctor / dtor bodies to help compiler optimizer
The compiler will generate both bodies of special functions and
deduce 'constexpr' and 'noexcept' specifiers for them automatically.

Note that when you move special functions from the class body, the
compiler considers them user-defined.  This might lead to
pessimization, so it's best to add '= default' directly in the class body
if possible.